Fix port in contact list server property
git-svn-id: http://kopete-xfire.googlecode.com/svn/trunk@35 8f771f7a-eddb-bb98-c32b-a9e201bdf2a8

+++ src/xf_account.cpp
@@ -233,7 +233,7 @@ void XfireAccount::updateContactSID(Xfire::Int32Attribute *p_id, Xfire::SIDAttri
     }
 }
 
-void XfireAccount::updateContactGameInformation(const Xfire::SessionID &p_sid, quint32 p_gameId, quint32 p_serverIp, quint32 p_serverPort)
+void XfireAccount::updateContactGameInformation(const Xfire::SessionID &p_sid, quint32 p_gameId, quint32 p_serverIp, quint16 p_serverPort)
 {
     XfireContact *c = static_cast<XfireContact*>(findContact(p_sid));
     if(c)
